Fix build

+++ src/coreclr/jit/morph.cpp
@@ -9120,9 +9120,9 @@ GenTree* Compiler::fgMorphCopyStruct(GenTreeOp* asg)
                 if (GenTreeIntCon* offset = addr->AsOp()->GetOp(1)->IsIntCon())
                 {
 #ifdef TARGET_XARCH
-                    unsigned maxOffset = INT32_MAX;
+                    ssize_t maxOffset = INT32_MAX;
 #else
-                    unsigned maxOffset = 4095;
+                    ssize_t maxOffset = 4095;
 #endif
                     // The maximum offset depends on the last promoted field offset, not the struct
                     // size. But such cases are so rare that it's not worth the trouble to get the
